pagemap: explicitly use virtual addresses
I am deeply unhappy about the use of size_t

+++ src/mem/pagemap.h
@@ -183,7 +183,7 @@ namespace snmalloc
           return std::pair(nullptr, 0);
 
         shift -= BITS_PER_INDEX_LEVEL;
-        ix = (addr >> shift) & ENTRIES_MASK;
+        ix = ((size_t)addr >> shift) & ENTRIES_MASK;
         e = &value->entries[ix];
 
         if constexpr (INDEX_LEVELS == 1)
@@ -203,7 +203,7 @@ namespace snmalloc
         return std::pair(nullptr, 0);
 
       shift -= BITS_FOR_LEAF;
-      ix = (addr >> shift) & LEAF_MASK;
+      ix = ((size_t)addr >> shift) & LEAF_MASK;
       return std::pair(leaf, ix);
     }
